Completely rewrote the way the program checks if a given answer is correct, such that score is no longer rewarded if two different questions share the same answers.
in doing so, I also fixed being able to select more than one answer per row, since now selected answers is represented as an object where the key is the text of the question being selected and the value is the answer the user selected. Therefore, only one answer can be chosen per question since it overwrites the old value found at the key, compared to the old method of just stuffing it into a limited size array.